Extracted generic `<GhTagsTokenInput>` out of `<GhPsmTagsInput>`
ref https://linear.app/ghost/issue/PROD-1605/

- tag searching & infinite scroll behaviour was built into the `<GhPsmTagsInput>` component but that had some ties to it's specific use-case with using `@post.tags` and saving the post
- we have a re-implementation of the tags token input in our "Add tag" bulk-edit modal which can't use the `<GhPsmTagsInput>` due to those specific use-case ties
- added a new re-usable `<GhTagsTokenInput>` component with the guts extracted from `<GhPsmTagsInput>` ready for use in the bulk add-tag modal
