+++ ghost/admin/app/components/gh-tags-token-input.hbs
@@ -1,0 +1,22 @@
+<GhTokenInput
+    @extra={{hash
+        tokenComponent=(component "gh-token-input/tag-token")
+    }}
+    @onChange={{this.updateTags}}
+    @onCreate={{this.createTag}}
+    @onOpen={{this.loadInitialTags}}
+    @options={{this.availableTags}}
+    @optionsComponent={{component "power-select-vertical-collection-options" lastReached=(perform this.loadMoreTagsTask)}}
+    @registerAPI={{this.registerPowerSelectAPI}}
+    @renderInPlace={{@renderInPlace}}
+    @search={{perform this.searchTagsTask}}
+    @selected={{@selected}}
+    @showCreateWhen={{this.showCreateWhen}}
+    @triggerId={{@triggerId}}
+    @triggerClass={{@triggerClass}}
+    @placeholder={{@placeholder}}
+    @allowCreation={{@allowCreation}}
+    @disabled={{@disabled}}
+    @selectedItemComponent={{@selectedItemComponent}}
+    @class={{@class}}
+/>
